Project now has ability to create an executable jar
