+++ pom.xml
@@ -16,16 +16,14 @@
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
-
-        <!--    Child dependency versions    -->
-        <javafx.maven.version>11</javafx.maven.version>
-        <javafx.maven.plugin.version>0.0.8</javafx.maven.plugin.version>
-        <maven.surefire.plugin.version>3.0.0-M5</maven.surefire.plugin.version>
-
-        <!--    Child module versions    -->
-        <game-logger-backend.version>0.0.1-SNAPSHOT</game-logger-backend.version>
-        <game-logger-ui.version>0.0.1-SNAPSHOT</game-logger-ui.version>
-        <game-logger-assembly.version>0.0.1-SNAPSHOT</game-logger-assembly.version>
     </properties>
+
+    <distributionManagement>
+        <repository>
+            <id>github</id>
+            <name>GitHub carpconn Apache Maven Packages</name>
+            <url>https://maven.pkg.github.com/carpconn/game-logger</url>
+        </repository>
+    </distributionManagement>
 
 </project>
